+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to the "ev3dev-browser" extension will be documented in this
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
 ## Unreleased
+### Added
+- Device connect/disconnect commands
+- Command to capture a screenshot from the remote device
 ### Fixed
 - Fix downloading projects with subdirectories
-### Added
-- Command to capture a screenshot from the remote device
 
 ## 0.2.0 - 2017-08-15
 ### Added